requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==1.10.13
email-validator==2.1.0.post1
jinja2==3.1.2
psycopg2-binary
pyjwt==2.8.0
pytest==7.4.3
python-multipart==0.0.6